changing yaml libraries to the k8s one which has better cross support
